+++ main.tf
@@ -77,6 +77,10 @@ resource "google_container_cluster" "primary" {
     }
   }
 
+  workload_identity_config {
+    workload_pool = "${var.project_id}.svc.id.goog"
+  }
+
   addons_config {
     horizontal_pod_autoscaling {
       disabled = var.horizontal_pod_autoscaling_disabled
@@ -94,7 +98,7 @@ resource "google_container_cluster" "primary" {
 
 
 resource "google_container_node_pool" "primary_preemptible_nodes" {
-  name     = "preempt-pool"
+  name     = var.gke_nodepool_name
   location = var.regional ? var.region : var.zone
   cluster  = google_container_cluster.primary.name
 
@@ -119,7 +123,6 @@ resource "google_container_node_pool" "primary_preemptible_nodes" {
       disable-legacy-endpoints = "true"
     }
 
-
     service_account = google_service_account.gke_sa.email
     oauth_scopes    = var.oauth_scopes
 
@@ -130,6 +133,12 @@ resource "google_container_node_pool" "primary_preemptible_nodes" {
         value  = taint.value["value"]
         effect = taint.value["effect"]
       }
+    }
+
+    # https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/container_cluster#mode
+    # https://cloud.google.com/kubernetes-engine/docs/how-to/workload-identity#option_2_node_pool_modification
+    workload_metadata_config {
+      mode = var.workload_metadata_enabled ? "GKE_METADATA" : "GCE_METADATA"
     }
   }
 }