+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2022-06-02
+### Added
+- New variable `gke_nodepool_name` (default: `preempt-pool`)
+- Workload Identity Pool (enabled on cluster, but disabled on node-pool which can be enabled by setting `workload_metadata_enabled` to `true`)
+### Changed
+- Updated `examples/gke-gateway-controller` version from `v0.3.0` to `v0.4.3`
+
 ## [0.6.2] - 2022-05-17
 ### Added
 - Apache v2 license